+++ include/neuralnetwork/neuralnetwork.cpp
@@ -1237,17 +1237,28 @@ double NeuralNetwork::calculate_learning_rate(
     const int warmup_epochs = static_cast<int>(std::round(_options.learning_rate_warmup_target() * number_of_epoch));
     const int relative_epoch = epoch - warmup_epochs;
 
-    learning_rate = calculate_smooth_learning_rate_boost(relative_epoch, learning_rate, boost_interval, per_boost_ratio);
-
-    // are we decaying the learning rate?
-    // this is done after warmup
-    if (learning_rate_decay_rate != 0)
+    if (_options.cosine_annealing_warm_restarts().enabled())
     {
-      learning_rate = learning_rate * std::exp(-learning_rate_decay_rate * relative_epoch);
+      learning_rate = _options.cosine_annealing_warm_restarts().calculate_learning_rate(relative_epoch, learning_rate_base);
       Logger::trace([=] 
         {
-          return Logger::factory("Learning rate to ", std::fixed, std::setprecision(15), learning_rate, " at epoch ", epoch, " (", std::setprecision(4), (double)(completed_percent * 100.0), "%)");
+          return Logger::factory("Cosine annealing learning rate to ", std::fixed, std::setprecision(15), learning_rate, " at epoch ", epoch, " (", std::setprecision(4), (double)(completed_percent * 100.0), "%)");
         });
+    }
+    else
+    {
+      learning_rate = calculate_smooth_learning_rate_boost(relative_epoch, learning_rate, boost_interval, per_boost_ratio);
+
+      // are we decaying the learning rate?
+      // this is done after warmup
+      if (learning_rate_decay_rate != 0)
+      {
+        learning_rate = learning_rate * std::exp(-learning_rate_decay_rate * relative_epoch);
+        Logger::trace([=] 
+          {
+            return Logger::factory("Learning rate to ", std::fixed, std::setprecision(15), learning_rate, " at epoch ", epoch, " (", std::setprecision(4), (double)(completed_percent * 100.0), "%)");
+          });
+      }
     }
   }